backend/worker/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.32.3
